Moved project membership creation to project membership fns
